requirements.txt
requests
pytest
structlog
curlify
pydantic
pyhamcrest
psycopg2
records
sqlalchemy